+++ CHANGELOG.md
@@ -2,12 +2,17 @@
 
 ## Current released version
 
+### 1.0.0 RC9 - 2014-06-21
+
+- Added jekyll-page-hooks as a dependency.
+- Now processing partials as ConvertiblePartials
+
+## Past versions
+
 ### 1.0.0 RC8 - 2014-06-09
 
 - Improved the way Ink manages the payload
 - Added `linkposts` and `articles` methods to Ink
-
-## Past versions
 
 ### 1.0.0 RC7 - 2014-06-08